pyproject.toml
[build-system]
requires = ["setuptools>=61.00", "setuptools_scm[toml]>=5", "wheel", "pytest"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_scheme = "no-guess-dev"
tag_regex = '''^(?:[\w-]+)?(?P<version>[vV]?\d+(?:\.\d+){0,2}[^\+]*)(?:\+.*)?$'''

[tool.setuptools.packages.find]
where = ["backend"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests/backend",
]

[project]
name = "DPKI"
dynamic = ["version"]
description = "Distributed public key infrastructure"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytoml>=0.1.21",
    "sqlalchemy>=2.0.3",
    "cryptography>=39.0.1",
    "aiosqlite>=0.18.0",
    "httpx>=0.23.3",
    "PyTend-ABCI @  git+https://github.com/curtapp/PyTend-ABCI.git@master#egg=PyTend-ABCI",
]


[project.scripts]
testnet-gen = "scripts.testnet:main"
request-cert = "scripts.x509gen:main"
